pyproject.toml
[project]
name = "fastapi-celery-notificationservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi==25.1.0",
    "asyncpg==0.30.0",
    "celery==5.5.3",
    "debugpy==1.8.16",
    "fastapi[standard]==0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings==2.10.1",
    "pyjwt>=2.10.1",
    "tortoise-orm==0.25.1",
]
